Removal of files no longer needed. New example of field-level vsoutjection. Properties file added for properties resolution. ID list as an Integer. This is now a Long. Corrected in User and Dao class.

+++ src/main/java/com/allanshoulders/vsoutjection/sample/dao/UserRowMapper.java
@@ -16,7 +16,7 @@ public class UserRowMapper implements ParameterizedRowMapper<User> {
 		user.setFirstName(resultSet.getString(UsersColumns.FIRST_NAME.toString()));
 		user.setLastName(resultSet.getString(UsersColumns.LAST_NAME.toString()));
 		user.setMiddleName(resultSet.getString(UsersColumns.MIDDLE_NAME.toString()));
-		user.setId(resultSet.getInt(UsersColumns.ID.toString()));
+		user.setId(resultSet.getLong(UsersColumns.ID.toString()));
 		user.setCity(resultSet.getString(UsersColumns.CITY.toString()));
 		user.setState(resultSet.getString(UsersColumns.STATE.toString()));
 		user.setTwitterName(resultSet.getString(UsersColumns.TWITTER_NAME.toString()));
